requirements.txt
python-dotenv~=1.1.1
pymssql~=2.3.8
streamlit~=1.50.0
pandas~=2.3.3
altair~=5.5.0
tqdm~=4.67.1
langgraph~=1.0.1
langchain
langchain-openai~=1.0.1
sqlalchemy~=2.0.44
fastapi~=0.120.0
uvicorn
pydantic~=2.12.3
langchain-core~=1.0.1
requests~=2.32.3
htbuilder~=0.7.0